Boss Dragon Fogwall
Implements a fogwall for the void dragon boss. Increases the health by a good bit. Has it drop a key to dissipate the fogwall. Removes three of the six(6!!!!) dragon rings. Zezuz Pyst. Places unbreakable walls around the surrounding areas to assure you can't just mine in and fire at it from the north, either.